% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.basic_tools.R
\name{across_df_TF}
\alias{across_df_TF}
\title{Across a df to count TRUE and FALSE}
\usage{
across_df_TF(df, type = "T")
}
\arguments{
\item{df}{a data frame}

\item{type}{"T" for TRUE counts (default), "F" for FALSE counts}
}
\value{
a data frame with the number of TRUE or FALSE values in each column
}
\description{
This function summarizes both TRUE and FALSE values in each column of a data frame.
}
\examples{
df <- data.frame(a = c(TRUE, FALSE, TRUE, TRUE), b = c(FALSE, TRUE, TRUE, TRUE))
across_df_TF(df) # Count TRUE (default)
across_df_TF(df, "F") # Count FALSE
}
