% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre.R
\name{add_rsid}
\alias{add_rsid}
\title{Convert CHR:BP to rsID}
\usage{
add_rsid(dat, ref = "GRCh37")
}
\arguments{
\item{ref}{str indicating ref version.
\itemize{
\item "37" for GRCh37
\item "38" for GRCh38
}}

\item{df}{A dataframe containing at least the columns 'CHR' and 'BP'.}
}
\value{
A dataframe with an additional 'RefSNP_id' column which contains the rsIDs.
}
\description{
This function takes a dataframe that must include columns 'CHR' and 'BP',
and it appends the corresponding rsID by querying the SNPlocs.Hsapiens.dbSNP155.GRCh37 database.
The function returns a dataframe with the rsIDs included.
}
\examples{
pacman::p_load(data.table, BSgenome, leo.gwas)
df <- data.frame(CHR = c(1, 1), BP = c(15211, 15820))
result <- add_rsid(df); result
}
