% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.annotation.R
\name{annotate_cpg_sites}
\alias{annotate_cpg_sites}
\title{Annotate CpG Sites with Gene Information}
\usage{
annotate_cpg_sites(
  cpg_vector,
  annotation_package = "IlluminaHumanMethylation450kanno.ilmn12.hg19"
)
}
\arguments{
\item{cpg_vector}{A character vector of CpG site probe IDs to be annotated.}

\item{annotation_package}{A character string specifying the Illumina annotation package to use.
It can be one of the following:
\itemize{
\item "IlluminaHumanMethylation450kanno.ilmn12.hg19" (default)
\item "IlluminaHumanMethylationEPICanno.ilm10b4.hg19"
}}
}
\value{
A data frame with two columns:
\itemize{
\item \code{CpG_Site}: The original CpG site probe IDs.
\item \code{Gene}: The associated gene names. \code{NA} if no gene annotation is found.
}
}
\description{
This function annotates a vector of CpG site probe IDs by retrieving corresponding gene names
from the Illumina 450k annotation package. It returns a data frame with the original CpG sites
and their associated gene names.
}
\examples{
\dontrun{
library(IlluminaHumanMethylation450kanno.ilmn12.hg19)
library(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
# Example CpG probe IDs
cpg_ids <- c("cg00000029", "cg00000108", "cg00000109")
# Annotate CpG sites
annotate_cpg_sites(cpg_ids, "IlluminaHumanMethylation450kanno.ilmn12.hg19")
}
}
