% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.visual_tools.R
\name{blank_plot}
\alias{blank_plot}
\title{Generate a blank plot with a message}
\usage{
blank_plot(message)
}
\arguments{
\item{message}{Text to display on the blank plot}
}
\value{
A ggplot object
}
\description{
Generate a blank plot with a message
}
\keyword{internal}
