% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_packages.R
\name{check_significant_SNP}
\alias{check_significant_SNP}
\title{locate the significant SNP for conditional analysis}
\usage{
check_significant_SNP(x, environment.list, significance_level = 5e-08)
}
\arguments{
\item{x}{data.frame of the SNP information}

\item{environment.list}{tibble of the environment list (see the example for usage)}

\item{significance_level}{numeric, significance level, default is 5e-8}
}
\value{
message that informs the user of the independant SNP for the following conditional analysis
}
\description{
locate the significant SNP for conditional analysis
}
\section{HLA data analysis}{

}

\examples{
con_dir <- "/Users/leoarrow/project/VKH2024/data/zuo/con_su" # specify the directory to store the HLA original data and subsequent conditional analysis results data
files <- list.files(con_dir,full.names = T) \%>\% as.vector(); files # update it each time
x1 <- fread(files[1]) \%>\% arrange(desc(CHISQ)) # read the data and sort it by CHISQ/P value
x2 <- fread(files[2]) \%>\% arrange(P) # repeat it until no more independent signal can be found.
x3 <- fread(files[3]) \%>\% arrange(P)
x4 <- fread(files[4]) \%>\% arrange(P)
x5 <- fread(files[5]) \%>\% arrange(P)
x6 <- fread(files[6]) \%>\% arrange(P)
x7 <- fread(files[7]) \%>\% arrange(P)
x8 <- fread(files[8]) \%>\% arrange(P)

env <- ls() # get the environment; this line if were put in main func will lead to error.
environment.list <- tibble(item = as.vector(grep("^x[0-9]+$", x = env, value = TRUE))) # load the environment
check_significant_SNP(x8, environment.list, significance_level = 5e-8)

# You can mannually check the p-value of one SNP in previous environment.list

}
