% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.basic_tools.R
\name{chisq_p_value}
\alias{chisq_p_value}
\title{Give precise p-value for chi-square test}
\usage{
chisq_p_value(chisq_value, df, digits = 4, prec = 100)
}
\arguments{
\item{chisq_value}{numeric, chi-square value}

\item{df}{numeric, degree of freedom}

\item{digits}{numeric, digits for output to illustrate}

\item{prec}{numeric, precision for mpfr() function}
}
\value{
A precise p-value in scientific format
}
\description{
Sometimes you get a p-value of 0 when you perform a chi-square test or other analysis.
This is because the p-value is so small that R rounds it to 0. This function gives you
a more precise p-value.
}
\examples{
# install.packages("Rmpfr")
library(Rmpfr)
chisq_value <- 2629; df <- 1
p_value <- chisq_p_value(chisq_value, df)
print(p_value)
}
