% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_tools.R
\name{clump_data_local}
\alias{clump_data_local}
\title{Clump data locally}
\usage{
clump_data_local(dat, pop = NULL, bfile = NULL, plink_bin = NULL)
}
\arguments{
\item{dat}{a datafame with \code{SNP}, \code{pval.exposure} and \code{id.exposure} col}

\item{pop}{Super-population to use as reference panel. Options are: AFR, AMR, EAS, EUR, SAS}

\item{bfile}{If this is provided then will go ld locally}
}
\value{
a subset of input dat with independant SNP
}
\description{
Here is the reference:
https://github.com/MRCIEU/TwoSampleMR/issues/173
https://blog.csdn.net/xiaozheng1213/article/details/126269969
library(ieugwasr)
library(plinkbinr) # devtools::install_github("explodecomputer/plinkbinr")
plinkbinr::get_plink_exe()
}
\keyword{tsmr:}
