% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_besd_tools.R
\name{combine_smr_res_1outcome}
\alias{combine_smr_res_1outcome}
\title{Combine \code{.fdr} files for one or multiple outcomes (seperately)}
\usage{
combine_smr_res_1outcome(
  dir,
  outcome,
  out_dir = file.path(dir, "combine_1outcome")
)
}
\arguments{
\item{dir}{Character. The parent folder that contains subfolders with fdr files.}

\item{outcome}{Character or character vector. One or more outcomes to search in file names.}

\item{out_dir}{Character. Output folder; default is to create "combine_1outcome" under \code{dir}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     NULL. This function writes combined files to disk directly.
}\if{html}{\out{</div>}}
}
\description{
Combine \code{.fdr} files for one or multiple outcomes (seperately)
}
\examples{
\dontrun{
combine_smr_res_1outcome("/Users/leoarrow/project/iridocyclitis/output/smr", "iri3")
}
}
