% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.basic_tools.R
\name{count_duplicate_element}
\alias{count_duplicate_element}
\title{Count or Identify Duplicates in a Vector}
\usage{
count_duplicate_element(vector, return = "count")
}
\arguments{
\item{vector}{A vector to be checked for duplicates.}

\item{return}{A string specifying the return type: "count" for number of duplicates, or "logi" for a logical vector.}
}
\value{
An integer representing the number of duplicates if return is "count",
or a logical vector indicating which elements are duplicates if return is "logi".
}
\description{
This function counts the number of duplicate elements in a vector,
or returns a logical vector indicating which elements are duplicates.
}
\examples{
vec <- c("a", "b", "c", "a", "b", "d")
count_duplicate_element(vec, return = "count")  # Returns 4
count_duplicate_element(vec, return = "logi")   # Returns c(TRUE, TRUE, FALSE, TRUE, TRUE, FALSE)
}
