% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.basic_tools.R
\name{count_matching_elements}
\alias{count_matching_elements}
\title{Count or Identify Matches of a Pattern in a Vector}
\usage{
count_matching_elements(vector, pattern, return = "count")
}
\arguments{
\item{vector}{A character vector to be searched.}

\item{pattern}{The pattern to match (regular expression).}

\item{return}{A string specifying the return type: "count" for number of matches, or "logi" for a logical vector.}
}
\value{
An integer representing the number of matches if return is "count",
or a logical vector indicating which elements match the pattern if return is "logi".
}
\description{
This function counts the number of elements in a vector that contain a given pattern,
or returns a logical vector indicating which elements match the pattern.
}
\examples{
vec <- c("abc", "def", "xyz", "abcd")
count_matching_elements(vec, "abc", return = "count")  # Returns 2
count_matching_elements(vec, "abc", return = "logi")   # Returns c(TRUE, FALSE, FALSE, TRUE)
}
