% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs_tools.R
\name{dr.prs}
\alias{dr.prs}
\alias{combine_rank}
\title{DuoRank PRS (Dr.PRS)}
\usage{
dr.prs(
  stage1_bfile = "./data/zuo/bed/SNP_for_PRS/VKH-zhonghua-for_prs_snplist",
  stage2_bfile = "./data/zuo/bed/SNP_for_PRS/VKH-ASA-for_prs_snplist",
  summary_file = "./output/part1-gwas/prs/snp_for_prs.txt",
  output_dir = "./output/part1-gwas/prs/b1t2",
  snp_col = "SNP.meta",
  a1_col = "A1",
  weight_col = "OR_Final",
  weight_type = "OR",
  divide_ratio = 0.7,
  iLasso_iteration = 1000,
  plink_bin = plinkbinr::get_plink_exe(),
  seed = 725
)

combine_rank(rank1, rank2, auc1 = NULL, auc2 = NULL)
}
\arguments{
\item{stage1_bfile}{Path to the stage1 PLINK binary files. Normally it is the one generates summary data. If only 1 source of individual data is available, you can split it into 2 non-overlap datasets.}

\item{stage2_bfile}{Path to the stage2 PLINK binary files, i.e., the target dataset for PRS calculation.}

\item{summary_file}{Path to the summary statistics file or a data frame containing SNPs, alleles, and weights. Note this summary only contains refined SNP for PRS calculation.}

\item{output_dir}{Directory to save the output files.}

\item{snp_col}{Column name in the summary statistics for SNP IDs (default: "SNP.meta").}

\item{a1_col}{Column name in the summary statistics for effect allele (default: "A1").}

\item{weight_col}{Column name in the summary statistics for weights (default: "OR_Final").}

\item{weight_type}{Type of weights, either "OR" (default) or "Beta". If "OR", it will calculate Beta values.}

\item{plink_bin}{Path to the PLINK binary executable (default: \code{plinkbinr::get_plink_exe()}).}

\item{seed}{Random seed. Default: 725}
}
\value{
A list with:
\itemize{
\item lasso_importance, catboost_importance, combined_rank
\item greedy_auc_curve (data.frame: kept_snps, auc)
\item final_subset (character vector of SNPs)
\item plink_prs_file (path), logs
}
}
\description{
Dr.PRS is designed to rank the importance for PRS inputs and futher optimazation.
It takes two none-overlap stage plink file (bfile) for machine learning modeling:
\itemize{
\item lasso \link{to capture linear relationship}
\item catboost \link{to capture non-linear relationship}
It also calculated PRS with plink using traditional additive model.
}
}
