% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.basic_tools.R
\name{exclude_HLA}
\alias{exclude_HLA}
\title{Exclude HLA region from genomic summary data}
\usage{
exclude_HLA(
  data,
  chromosome_col = "CHR",
  position_col = "BP",
  lower_bound = 2.5e+07,
  upper_bound = 3.4e+07
)
}
\arguments{
\item{data}{A data frame containing genomic data.}

\item{chromosome_col}{The name of the column representing chromosome numbers (default is CHR).}

\item{position_col}{The name of the column representing genomic positions (default is BP).}

\item{lower_bound}{The lower boundary of the HLA region in base pairs (default is 25e6).}

\item{upper_bound}{The upper boundary of the HLA region in base pairs (default is 34e6).}
}
\value{
A data frame excluding rows that fall within the specified HLA region.
}
\description{
This function filters out entries within the HLA region on a specified chromosome
and position range. The default HLA region is set to chromosome 6, between 25Mb and 34Mb.
Custom chromosome and position bounds can be specified.
}
\examples{
example_data <- data.frame(
  SNP = c("rs1", "rs2", "rs3", "rs4"),
  CHR = c(6, 6, 6, 7),
  POS = c(26000000, 33000000, 35000000, 29000000)
)
result_data <- exclude_HLA(example_data, chromosome_col="CHR", position_col="POS")
print(result_data)
}
