% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_packages.R
\name{extract_instruments_local}
\alias{extract_instruments_local}
\title{Extract instruments locally}
\usage{
extract_instruments_local(
  dat,
  p = 5e-08,
  N = "Neff",
  pop = "EUR",
  bfile = NULL,
  plink_bin = NULL
)
}
\arguments{
\item{dat:}{Summary statistics needs: SNP, CHR, POS, A1, A2, EAF, BETA, SE, P, Phenotype, N}

\item{p:}{P cut-off value}

\item{pop:}{param for clump_data_local}

\item{bfile:}{param for clump_data_local}
}
\value{
a clumped tsmr format dataframe
}
\description{
Extract instruments locally
}
\details{
要求dat数据中P的列名为P
}
