% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_tools.R
\name{format_outcome}
\alias{format_outcome}
\title{format outcome data}
\usage{
format_outcome(dat, snp = iv$SNP, N = "Neff")
}
\arguments{
\item{dat}{a dataframe for outcome with SNP, CHR, POS, A1, A2, EAF, BETA, SE, P, Phenotype, samplesize columns}

\item{snp}{a str vector out of iv$SNP}

\item{N}{N column name for sample size (effective or observed)}
}
\value{
a tsmr format outcome dataframe
}
\description{
format outcome data
}
\keyword{tsmr:}
