% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre.R
\name{get_id}
\alias{get_id}
\title{Title Get ID using CHR, BP, A2 (REF/Non-effect), A1 (ALT/Effect)}
\usage{
get_id(x)
}
\arguments{
\item{x}{df data.frame with CHR, BP, A2, A1 columns}
}
\value{
A df data.frame with ID column and the character counts of A1 and A2
}
\description{
Title Get ID using CHR, BP, A2 (REF/Non-effect), A1 (ALT/Effect)
}
