% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.visual_tools.R
\name{group_comparison_draw}
\alias{group_comparison_draw}
\title{Draw Group Comparison for a Continuous Variable}
\usage{
group_comparison_draw(
  df,
  x_col,
  y_col,
  vector_x = NULL,
  vector_y = NULL,
  test_method = "wilcox",
  title = "Group Comparison",
  xlab = NULL,
  ylab = NULL,
  alpha = 0.8,
  main_color = "#BB7CD8",
  violin_fill = main_color,
  box_fill = main_color,
  box_color = "black",
  jitter = T,
  jitter_size = 2,
  jitter_color = "black",
  drop_na = F,
  annotate_n = T,
  title_size = 16,
  xlab_size = 14,
  ylab_size = 14,
  axis_text_size = 14
)
}
\arguments{
\item{df}{A data frame (optional if vectors provided).}

\item{x_col}{Name of categorical variable in df.}

\item{y_col}{Name of numeric variable in df.}

\item{vector_x}{Optional categorical vector.}

\item{vector_y}{Optional numeric vector.}

\item{test_method}{Statistical test: \code{"wilcox"} (default, Mann–Whitney U test) or \code{"t.test"}.}

\item{title}{Plot title.}

\item{xlab, ylab}{Axis labels.}

\item{alpha}{Transparency for violin, boxplot and jitter.}

\item{main_color}{Easy way to set the vibe. Good luck trying!}

\item{violin_fill}{Fill color for violin.}

\item{box_fill}{Fill color for boxplot.}

\item{box_color}{Outline color for boxplot.}

\item{jitter}{Logical, whether to show jitter. Default TRUE.}

\item{jitter_size}{Size of jitter points. Default \code{2}.}

\item{jitter_color}{Color of jitter points. Default \code{"black"}.}

\item{drop_na}{Logical, whether to drop NA group from x variable. Default \code{FALSE}.}

\item{annotate_n}{Logical, whether to annotate sample size N for each group. Default \code{TRUE}.}

\item{title_size}{Font size for plot title. Default \code{16}.}

\item{xlab_size}{Font size for x-axis label. Default \code{14}.}

\item{ylab_size}{Font size for y-axis label. Default \code{14}.}

\item{axis_text_size}{Font size for axis text. Default \code{14}.}
}
\value{
A ggplot2 object.
}
\description{
This function creates a violin + boxplot to visualize the distribution of a continuous variable
across groups of a categorical variable, with optional jitter points and annotated p-value.
}
\examples{
df <- tibble::tibble(group = rep(c(0,1), each=50), prs = rnorm(100))
group_comparison_draw(df, "group", "prs")
group_comparison_draw(df, "group", "prs", annotate_n = TRUE)
group_comparison_draw(vector_x = rep(c(0,1), each=50), vector_y = rnorm(100))
group_comparison_draw(vector_x = rep(c(0,1), each=50), vector_y = rnorm(100), jitter = FALSE)
}
