% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.visual_tools.R
\name{ld_ps_index}
\alias{ld_ps_index}
\title{Loci_plot: Calculate the LD-matrix (LD r2) for the index SNP}
\usage{
ld_ps_index(
  gwas,
  index = "rs999",
  win = 1000,
  ld_calculation = T,
  bfile = "/Users/leoarrow/project/ref/1kg.v3/EAS",
  plink_bin = plinkbinr::get_plink_exe()
)
}
\arguments{
\item{gwas}{gwas summary data that needs to select loci and calculate r2}

\item{index}{index snp}

\item{win}{window size to locally calculate the r2; set it larger than that you want to plot; # default calculate 1MB}

\item{ld_calculation}{if calculate the LD locally, defaut T; use F if the index snp is a rare variant (MAF<0.01)}

\item{bfile}{bfile}

\item{plink_bin}{plinkbinr::get_plink_exe()}

\item{pop}{only applicable under 500 snps; dont use it anyway}
}
\value{
loci data with calculated r2
}
\description{
Loci_plot: Calculate the LD-matrix (LD r2) for the index SNP
}
