% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.basic_tools.R
\name{leo_log}
\alias{leo_log}
\title{Log Messages with Timestamps}
\usage{
leo_log(
  ...,
  level = "info",
  levels = c("info", "success", "warning", "danger")
)
}
\arguments{
\item{...}{The message string to log, which will be pasted together.}

\item{level}{The log level. Options are \code{"info"}, \code{"success"}, \code{"warning"}, and \code{"danger"}.}

\item{levels}{All levels that is now supported.}
}
\value{
No return value. Outputs a formatted log message with a timestamp.
}
\description{
Logs messages with timestamps
The messages are styled using the \code{cli} package for enhanced readability.
This function can not deal with {} function in the \code{cli} package.
}
\examples{
leo_log("Processing the", n1, "and", n2, "files.")
leo_log("Task completed successfully!", level = "success")
leo_log("Potential issue detected.", level = "warning")
leo_log("Error occurred during processing!", level = "danger")
}
