% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.basic_tools.R
\name{leo_message}
\alias{leo_message}
\title{Give Messages with my color}
\usage{
leo_message(..., color = "31", return = FALSE)
}
\arguments{
\item{...}{The messages you wanna messgae, which will be pasted together}

\item{color}{Str. Preferred color. Default is yellow.
Options are "31" (red), "32" (green), "34" (blue), "95" (light purple)...}

\item{return}{Logical. If TRUE, returns the formatted string. If FALSE (Default), prints directly.}
}
\description{
Give Messages with my color
}
\examples{
leo_message("This is a red message", color = "31")
leo_message("This is a green message", "\nhaha", color = "32")
leo_message("This is a blue ", "message", color = "34")
leo_message("This is a light purple message", color = "95")
leo_message(" 游부游부游부 Welcome to use the LEO package ! 游부游부游부")
}
