% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.visual_tools.R
\name{leo_scale_color}
\alias{leo_scale_color}
\title{Apply Color Palette to ggplot}
\usage{
leo_scale_color(plot, color_palette = "npg")
}
\arguments{
\item{plot}{A ggplot object to which the color palette will be applied.}

\item{color_palette}{A character string specifying the color palette to use.
Options include \code{ggsci} palettes ("npg", "lancet", "jama", etc.),
\code{RColorBrewer} palettes, \code{viridis} palettes, or a custom vector of colors.}
}
\value{
A ggplot object with the applied color palette.
}
\description{
This function applies a specified color palette to a ggplot object,
supporting \code{ggsci}, \code{RColorBrewer}, and \code{viridis} palettes, as well as custom colors.
}
