% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_besd_tools.R
\name{leo_smr_adjust}
\alias{leo_smr_adjust}
\title{Adjust SMR Results with FDR and Bonferroni Corrections}
\usage{
leo_smr_adjust(smr_result_path, writePath = "", out_dir = "", write_out = T)
}
\arguments{
\item{smr_result_path}{Character. The path containing SMR result file. Files should have the \code{.smr} extension.}

\item{out_dir}{Character. The output directory where adjusted files will be saved.
If not specified, defaults to creat a dir named \code{fdr} within dirname(smr_result_path).}

\item{write_out}{Logical. Whether to write the adjusted results to a file. Default is \code{TRUE}.}
}
\description{
It applies FDR/Bonferroni corrections for a single SMR result.
The corrected results are saved in an \code{fdr} subdirectory within the output directory.
}
\examples{
\dontrun{
leo_smr_adjust("~/project/iridocyclitis/output/smr-t2d/sqtl/GTEx49/chr_combined/chr_combine_sQTL_Adipose_Subcutaneous@iri3.smr",
writePath = "",
out_dir = "~/project/iridocyclitis/output/smr-t2d/sqtl/GTEx49")
leo_smr_adjust("~/project/iridocyclitis/output/smr-t2d/sqtl/GTEx49/chr_combined/chr_combine_sQTL_Adipose_Subcutaneous@iri3.smr",
              writePath = "./haha.fdr",
              out_dir = "")
}
}
