% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.visual_tools.R
\name{locuszoomr_loc}
\alias{locuszoomr_loc}
\title{Loci_plot: prepare the locus data for locuszoomr}
\usage{
locuszoomr_loc(loci_data, gene, online_ld = F, index_snp, flank)
}
\arguments{
\item{loci_data}{output from ld_ps_index}

\item{gene}{gene loci}

\item{online_ld}{whether to use online LD; default is F}

\item{index_snp}{indexed snp}

\item{flank}{flank size for the locus plot}
}
\value{
prepared data which could be pass to save_regional_plot
}
\description{
Loci_plot: prepare the locus data for locuszoomr
}
