% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.annotation.R
\name{map_ensg_to_gene_using_org.Hs.eg.db}
\alias{map_ensg_to_gene_using_org.Hs.eg.db}
\title{Map Ensembl IDs to Gene Symbols using \code{org.Hs.eg.db}}
\usage{
map_ensg_to_gene_using_org.Hs.eg.db(ensembl_ids, ensembl_col = NULL)
}
\arguments{
\item{ensembl_ids}{A character vector of Ensembl IDs or a data frame containing Ensembl IDs.}

\item{ensembl_col}{The column name of Ensembl IDs if \code{ensembl_ids} is a data frame.}
}
\value{
A data frame with two columns: the input Ensembl IDs and the corresponding gene symbols.
}
\description{
This function maps Ensembl IDs to their corresponding gene symbols using the \code{org.Hs.eg.db} package.
}
\examples{
\dontrun{
ensembl_ids <- c("ENSG00000141510.1", "ENSG00000012048",
                 "ENSG00000146648") # Example with Ensembl ID vector
map_ensg_to_gene_using_org.Hs.eg.db(ensembl_ids = ensembl_ids)

ensembl_ids_df <- data.frame(EnsemblID = ensembl_ids,
                             OtherInformation = c(1,2,3)) # Example with data frame input
map_ensg_to_gene_using_org.Hs.eg.db(ensembl_ids = ensembl_ids_df,
                                      ensembl_col = "EnsemblID")
}
}
