% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.annotation.R
\name{map_gene_to_chrbp_using_TxDb}
\alias{map_gene_to_chrbp_using_TxDb}
\title{Map Gene Symbols Using Bioconductor Packages}
\usage{
map_gene_to_chrbp_using_TxDb(
  genes,
  gene_col = NULL,
  genome = c("hg19", "hg38")
)
}
\arguments{
\item{genes}{A character vector of gene symbols or a data frame containing gene symbols.}

\item{gene_col}{The column name of gene symbols if \code{genes} is a data frame.}

\item{genome}{The genome assembly to use: \code{"hg19"} or \code{"hg38"}.
\itemize{
\item For hg19, it needs \code{"TxDb.Hsapiens.UCSC.hg19.knownGene"} Bioconductor package
\item For hg38, it needs \code{"TxDb.Hsapiens.UCSC.hg38.knownGene"} Bioconductor package
}}
}
\value{
A data frame with mapped results.
}
\description{
Map Gene Symbols Using Bioconductor Packages
}
\examples{
\dontrun{
gene_symbols <- c("TP53", "BRCA1", "EGFR") # Example with gene symbol vector
map_gene_to_chrbp_using_TxDb(genes = gene_symbols, genome = "hg19")

gene_symbols_df <- data.frame(GeneName = gene_symbols, OtherInformation = c(1,2,3)) # Example with data frame input
map_gene_to_chrbp_using_TxDb(genes = gene_symbols_df, gene_col = "GeneName" , genome = "hg19")
}
}
