% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.annotation.R
\name{map_gene_to_tss_using_gtf}
\alias{map_gene_to_tss_using_gtf}
\title{Map Genes to Their TSS Positions}
\usage{
map_gene_to_tss_using_gtf(
  genes,
  gene_col = NULL,
  genome = c("hg19", "hg38"),
  gtf_file = NULL,
  download_dir = "~/project/ref/gtf",
  ...
)
}
\arguments{
\item{genes}{A character vector of gene symbols or a data frame containing gene symbols.}

\item{gene_col}{The column name of gene symbols if \code{genes} is a data frame.}

\item{genome}{The genome assembly to use: \code{"hg19"} or \code{"hg38"}.}

\item{gtf_file}{The path to a GTF file. If \code{NULL}, the function will download the appropriate GTF file.}

\item{download_dir}{The path where you want to store the downloaded gtf file.}

\item{...}{Pass to map_gene_to_chrbp_using_gtf. See \code{\link[=map_gene_to_chrbp_using_gtf]{map_gene_to_chrbp_using_gtf()}}}
}
\value{
A data frame with gene symbols and their TSS positions
}
\description{
This function maps gene symbols to their transcription start sites (TSS) positions
using hg19 or hg38 genome assembly.
}
\examples{
\dontrun{
genes <- c("TP53", "BRCA1", "EGFR")
map_gene_to_tss_using_gtf(genes = genes, genome = "hg38")

genes_df <- data.frame(GeneName = genes, OtherInfo = c(1,2,3))
map_gene_to_tss_using_gtf(genes = genes_df, gene_col = "GeneName", genome = "hg19")
}
}
