% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_tools.R
\name{mr_scatter_plot_modified}
\alias{mr_scatter_plot_modified}
\title{Modified MR Scatter Plot}
\usage{
mr_scatter_plot_modified(mr_results, dat)
}
\arguments{
\item{mr_results}{same as TwoSampleMR::mr_scatter_plot}

\item{dat}{same as TwoSampleMR::mr_scatter_plot}
}
\description{
The \code{mr_scatter_plot} in the \code{TwoSampleMR} package could be better.
This is a modified version of \code{mr_scatter_plot} in the \code{TwoSampleMR} package
When the slope of two method is really close, the scatter plot may not properly plot them!
Change the line type here mannually herein then.
}
\examples{
p1 <- mr_scatter_plot_modified(mr_results = res_pair, dat = dat_h_pair)
print(p1[[1]])
library(ggplot2);library(ggsci);library(TwoSampleMR);library(dplyr)
}
\keyword{tsmr:}
