% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.annotation.R
\name{report_mapping_stats}
\alias{report_mapping_stats}
\title{Report mapping statistics}
\usage{
report_mapping_stats(ensembl_col)
}
\description{
Report mapping statistics
}
\keyword{internal}
