% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.visual_tools.R
\name{save_regional_plot}
\alias{save_regional_plot}
\title{Loci_plot: save_regional_plot}
\usage{
save_regional_plot(
  path,
  loc,
  gene,
  save = T,
  title = expression(paste(italic("CLPSL1"), " (T1D)")),
  labels = c("index"),
  filter_gene_biotype = c("protein_coding"),
  border = F,
  width = 7.5,
  height = 5.5
)
}
\arguments{
\item{path}{Path to the directory containing summary statistics.}

\item{loc}{Locus string (e.g., "1:1000-2000").}

\item{gene}{Gene name to highlight.}

\item{save}{Logical. Whether to save the plot (default: TRUE).}

\item{title}{Title of the plot.}

\item{labels}{Labels to indicate index SNP and other SNPs.}

\item{filter_gene_biotype}{Character vector of gene biotypes to filter.}

\item{border}{Logical. Whether to add a border for gene track.}

\item{width}{Plot width.}

\item{height}{Plot height.}
}
\description{
Loci_plot: save_regional_plot
}
