% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_tools.R
\name{clump_data_local}
\alias{clump_data_local}
\title{Perform LD Clumping Locally or via Reference Panel}
\usage{
clump_data_local(
  dat,
  pop = NULL,
  bfile = NULL,
  clump_kb = 10000,
  clump_r2 = 0.001,
  plink_bin = plinkbinr::get_plink_exe()
)
}
\arguments{
\item{dat}{Data frame with columns \code{SNP}, \code{pval.exposure}; optional \code{id.exposure}.}

\item{pop}{1000G super-pop for online clumping (AFR/AMR/EAS/EUR/SAS). Used when \code{bfile} is NULL.}

\item{bfile}{PLINK reference panel prefix for local clumping (without .bed/.bim/.fam). If set, \code{pop} is ignored.}

\item{clump_kb}{Window size in kb. Default 10000.}

\item{clump_r2}{r^2 threshold. Default 0.001.}

\item{plink_bin}{Path to PLINK executable (auto-detect via plinkbinr if NULL and \code{bfile} is set).}
}
\description{
This function performs LD clumping using either a local PLINK binary file (\code{bfile})
or a 1000 Genomes super population panel. Requires the \code{ieugwasr} and \code{plinkbinr} packages.
}
\examples{
# Reference:
# - https://github.com/MRCIEU/TwoSampleMR/issues/173
# - https://blog.csdn.net/xiaozheng1213/article/details/126269969
library(ieugwasr)
library(plinkbinr) # devtools::install_github("explodecomputer/plinkbinr")
plinkbinr::get_plink_exe()
}
\keyword{tsmr:}
