% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.visual_tools.R
\name{correlation_draw}
\alias{correlation_draw}
\title{Draw Correlation between Two Vectors}
\usage{
correlation_draw(
  vector_x,
  vector_y,
  method = "spearman",
  color_palette = "npg",
  title = "Correlation Plot",
  xlab = "Vector X",
  ylab = "Vector Y",
  point_size = 1.5,
  point_color = "#BB7CD8",
  point_stroke = 1,
  alpha = 0.75,
  line_color = "#BB7CD8",
  line_type = "dashed",
  line_size = 1.2,
  ci_alpha = 0.2,
  title_size = 16,
  xlab_size = 14,
  ylab_size = 14,
  axis_text_size = 14,
  ...
)
}
\arguments{
\item{vector_x}{A numeric vector.}

\item{vector_y}{A numeric vector of the same length as \code{vector_x}.}

\item{method}{A character string specifying the correlation method ("spearman" or "pearson").
Defaults to "spearman".}

\item{color_palette}{A character string or vector specifying the color palette to use.
Can be a palette name from \code{ggsci}, \code{RColorBrewer}, or a custom color vector.}

\item{title}{A character string for the plot title. Defaults to "Correlation Plot".}

\item{xlab}{A character string for the x-axis label. Defaults to "Vector X".}

\item{ylab}{A character string for the y-axis label. Defaults to "Vector Y".}

\item{point_size}{Numeric value for point size in the scatter plot. Defaults to 4.}

\item{point_color}{A character string specifying color for the scatter points. Defaults to "#BB7CD8".}

\item{point_stroke}{It could also be NA!!!}

\item{alpha}{Numeric value for the transparency of points. Defaults to 0.75.}

\item{line_color}{A character string specifying color for the trend line. Defaults to "#BB7CD8".}

\item{line_type}{Character specifying the type of line ("solid", "dashed", etc.). Defaults to "dashed".}

\item{line_size}{Numeric value specifying the thickness of the trend line. Defaults to 1.2.}

\item{ci_alpha}{Numeric value for the transparency level of the confidence interval. Defaults to 0.6.}

\item{title_size}{Numeric value for title text size. Defaults to 14.}

\item{xlab_size}{Numeric value for x-axis label text size. Defaults to 12.}

\item{ylab_size}{Numeric value for y-axis label text size. Defaults to 12.}

\item{axis_text_size}{Numeric value for axis text size. Defaults to 10.}

\item{...}{Additional arguments passed to \code{\link[stats]{cor.test}}.}
}
\value{
A ggplot object representing the scatter plot with correlation information.
}
\description{
This function creates a scatter plot to visualize the correlation between two vectors,
displaying the correlation coefficient and p-value on the plot.
It uses \code{ggplot2} for visualization and \code{ggsci} for color schemes.
}
\examples{
vector_x <- c(10, 2, 3, 4, 5)
vector_y <- c(5, 6, 7, 8, 7)
correlation_draw(vector_x, vector_y, method = "pearson", point_size = 10, color_palette = "npg")
}
\seealso{
\code{\link{correlation_calculate}} for calculating correlation coefficients and p-values.

\code{\link[=leo_scale_color]{leo_scale_color()}} for applying color palettes to ggplot objects.
}
