% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_packages.R
\name{leo_message}
\alias{leo_message}
\title{Give Messages with my color}
\usage{
leo_message(msg, color = "31")
}
\arguments{
\item{msg}{messages to show; it should be a word pasted pharse.}

\item{color}{prefered color}
}
\description{
Give Messages with my color
}
\examples{
leo_message("This is a pink message")
leo_message("This is a green message","32")
leo_message("This is a blue message","34")
leo_message("This is a light purple message","95")
}
\keyword{basic:}
