% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.basic_tools.R
\name{leo_message}
\alias{leo_message}
\title{Give Messages with my color}
\usage{
leo_message(..., color = "31")
}
\arguments{
\item{...}{The messages you wanna messgae, which will be pasted together}

\item{color}{prefered color}
}
\description{
Give Messages with my color
}
\examples{
leo_message("This is a pink message")
leo_message("This is a green message", "\nhaha", color = "32")
leo_message("This is a blue ", "message", color = "34")
leo_message("This is a light purple message", color = "95")
}
