% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_besd_tools.R
\name{leo_smr_adjust}
\alias{leo_smr_adjust}
\title{Adjust SMR Results with FDR and Bonferroni Corrections}
\usage{
leo_smr_adjust(
  smr_result_path,
  writePath = "",
  out_dir = "",
  QTL_type = "",
  Source = "",
  Tissue = "",
  Outcome_name = "",
  add_info_cols = T,
  drop_non_heidi = T,
  hla = F,
  write_out = T
)
}
\arguments{
\item{smr_result_path}{Character. The path containing SMR result file. Files should have the \code{.smr} extension.}

\item{writePath}{\subsection{Character. The path to write the adjusted results.}{
}}

\item{out_dir}{Character. The output directory where adjusted files will be saved.
If not specified, defaults to create a dir named \code{fdr} within dir-name (smr_result_path).}

\item{QTL_type}{Character. The type of the QTL for SMR analysis.}

\item{Source}{Character. The name of the Source.}

\item{Tissue}{Character. The name of the Tissue.}

\item{Outcome_name}{\subsection{Character. The name of the Outcome.}{
}}

\item{add_info_cols}{Logical. Whether to add information columns for: QTL_type, Source, Tissue and Outcome. Default is \code{TRUE}.}

\item{drop_non_heidi}{Logical. Whether to drop Probes without HEIDI test information. Default is \code{TRUE}.}

\item{hla}{Logical. Whether to pre-exclude the HLA region probes. Default is \code{False}.}

\item{write_out}{Logical. Whether to write the adjusted results to a file. Default is \code{TRUE}.}
}
\description{
It applies FDR/Bonferroni corrections for a single SMR result.
The corrected results are saved in an \code{fdr} subdirectory within the output directory.
}
\examples{
\dontrun{
leo_smr_adjust("~/project/iridocyclitis/output/smr-t2d/sqtl/GTEx49/chr_combined/chr_combine_sQTL_Adipose_Subcutaneous@iri3.smr",
               writePath = "", out_dir = "~/project/iridocyclitis/output/smr-t2d/sqtl/GTEx49")
leo_smr_adjust("~/project/iridocyclitis/output/smr-t2d/sqtl/GTEx49/chr_combined/chr_combine_sQTL_Adipose_Subcutaneous@iri3.smr",
               writePath = "./haha.fdr", out_dir = "")
}
}
