% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.annotation.R
\name{map_gene_class_using_annotables}
\alias{map_gene_class_using_annotables}
\title{Map Gene Symbols to biotype & description via \strong{annotables}}
\usage{
map_gene_class_using_annotables(genes, gene_col = "Gene", quiet = FALSE)
}
\arguments{
\item{genes}{Character vector of gene symbols, or data frame/tibble containing gene symbols}

\item{gene_col}{Column name (when \code{genes} is a table), default \code{"Gene"}}

\item{quiet}{Logical; if \code{TRUE}, suppress progress messages}
}
\value{
Data frame with same structure as input, plus \code{biotype}, \code{description}, \code{infer_version}
}
\description{
Use local dataframes \code{annotables::grch38} and \code{annotables::grch37}
to add \emph{biotype} and \emph{description} to gene symbols, providing a source column
\code{infer_version} ("grch38" / "grch37" / "unmapped").
}
\details{
Logic:
\enumerate{
\item Left join with \strong{grch38}; if matched, \code{infer_version = "grch38"}
\item If not matched, fill with \strong{grch37}; if matched, \code{infer_version = "grch37"}
\item If still not matched, fill columns with placeholders as described above
}
}
\examples{
\dontrun{
map_gene_class_using_annotables(c("TP53","BRCA1","C14orf37"))

df <- tibble::tibble(Gene = c("TP53","XIST","C14orf37"))
map_gene_class_using_annotables(df, gene_col = "Gene")
}
}
