% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_packages.R
\name{mr_one_pair}
\alias{mr_one_pair}
\title{perform_mr_for_one_pair}
\usage{
mr_one_pair(
  dat_h,
  exp = "",
  out = "",
  save_plot = T,
  res_dir = "./output/tsmr",
  fig_dir = "./figure/tsmr"
)
}
\arguments{
\item{dat_h}{harmonized data}

\item{exp}{a str indicating the exposure in dat_h}

\item{out}{a str indicating the outcome in dat_h}

\item{save_plot}{only save the plot if T, defaut T.}

\item{res_dir}{dir path where the result fo MR analysis stored}

\item{fig_dir}{dir path where the figure of MR analysis stored}
}
\value{
list(res_pair=res_pair, res_pair_presso=res_pair_presso)
}
\description{
\code{perform_mr_for_one_pair} perform a MR for one pair of exp and out
}
\section{tsmr}{

}

\examples{
clusterEvalQ(cl, {
library(vroom)
library(tidyverse)
library(TwoSampleMR)
library(ggplot2)
library(ggsci)
source("./code/1.0.tsmr_packages.R")
})
clusterExport(cl, varlist = c("uni_pair","dat_h"))
>>>>>>>>>>>>>  draw and save draw   >>>>>>>>>>>>>
p1: Scatter plot
p2: Forest plot
p3: Funnel plot
p4: LOO plot
For Forest: width = 7, height = 8 (T1D)
For Forest: width = 7, height = 10 (T2D)
For others: width = 7, height = 6
fig_dir <- "./figure/tsmr"
}
