% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsmr_tools.R
\name{mrlap_one_pair}
\alias{mrlap_one_pair}
\title{mrlap_one_pair}
\usage{
mrlap_one_pair(
  exposure_data,
  outcome_data,
  exposure_name,
  outcome_name,
  ld_path,
  hm3_path,
  log = F,
  log_path = "."
)
}
\arguments{
\item{exposure_data}{exposure_data}

\item{outcome_data}{outcome_data}

\item{exposure_name}{exposure_name}

\item{outcome_name}{outcome_name}

\item{ld_path}{ldsc required file}

\item{hm3_path}{ldsc required file; tutorial use nomhc version list}

\item{log}{logi, if T, would save ldsc log to log_path, defaut F}

\item{log_path}{path to where you wanna store the ldsc log}
}
\description{
\code{mrlap_one_pair} perform the MR-lap for one pair of exposure and outcome
\href{https://github.com/n-mounier/MRlap?tab=readme-ov-file}{MR-lap}
}
\keyword{tsmr:}
