% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.visual_tools.R
\name{save_regional_plot}
\alias{save_regional_plot}
\title{Loci_plot: save_regional_plot}
\usage{
save_regional_plot(
  path,
  loc,
  gene,
  save = T,
  title = expression(paste(italic("CLPSL1"), " (T1D)")),
  labels = c("index"),
  filter_gene_biotype = c("protein_coding"),
  border = F,
  width = 7.5,
  height = 5.5
)
}
\arguments{
\item{path}{path to store the plot; make sure the path is exist}

\item{loc}{output from locuszoomr_loc}

\item{gene}{gene}

\item{save}{if T, will save plot to path; if F, return the plot only}

\item{labels}{labels; in case you need to indicate the index SNP and other SNP}

\item{border}{border for gene track}

\item{width}{width}

\item{height}{height}
}
\description{
Loci_plot: save_regional_plot
}
